+++ src/__main__.py
@@ -12,7 +12,7 @@ from . import config
 from . import tool_dashboard
 from . import tool_rpccurlhelper
 from . import tool_currencyconvert
-from . import tool_miningcalcs
+from . import tool_miningcalc
 from . import tool_opreturn
 from . import tool_braiinspool
 from . import tool_settings
@@ -72,7 +72,7 @@ if __name__ == "__main__":
     app.add_url_rule('/dashboard', 'dashboard', pywebio.platform.flask.webio_view( tool_dashboard.main ), methods=['GET', 'POST', 'OPTIONS'])  # need GET,POST and OPTIONS methods
     app.add_url_rule('/curl_formatter', 'curl_formatter', pywebio.platform.flask.webio_view( tool_rpccurlhelper.main ), methods=['GET', 'POST', 'OPTIONS'])  # need GET,POST and OPTIONS methods
     app.add_url_rule('/currency_converter', 'currency_converter', pywebio.platform.flask.webio_view( tool_currencyconvert.main ), methods=['GET', 'POST', 'OPTIONS'])  # need GET,POST and OPTIONS methods
-    app.add_url_rule('/mining_calcs', 'mining_calcs', pywebio.platform.flask.webio_view( tool_miningcalcs.main ), methods=['GET', 'POST', 'OPTIONS'])  # need GET,POST and OPTIONS methods
+    app.add_url_rule('/mining_calcs', 'mining_calcs', pywebio.platform.flask.webio_view( tool_miningcalc.main ), methods=['GET', 'POST', 'OPTIONS'])  # need GET,POST and OPTIONS methods
     app.add_url_rule('/opreturn', 'opreturn', pywebio.platform.flask.webio_view( tool_opreturn.main ), methods=['GET', 'POST', 'OPTIONS'])  # need GET,POST and OPTIONS methods
     app.add_url_rule('/braiinspool', 'braiinspool', pywebio.platform.flask.webio_view( tool_braiinspool.main ), methods=['GET', 'POST', 'OPTIONS'])  # need GET,POST and OPTIONS methods
     app.add_url_rule('/blockstreamsat', 'blockstreamsat', pywebio.platform.flask.webio_view( tool_blockstreamsat.main ), methods=['GET', 'POST', 'OPTIONS'])  # need GET,POST and OPTIONS methods
