+++ src/tool_miningcalc/calculator-src/config.py
