+++ src/tool_miningcalc/calculator-src/constants.py
