+++ src/tool_miningcalc/calculator-src/data.py
