+++ src/tool_miningcalc/calculator-src/run_calculator.sh
