+++ src/tool_miningcalc/calculator-src/tool_instant_calcs.py
