+++ src/tool_miningcalc/config.py
