+++ src/tool_treasurymanagement/config.py
@@ -1,20 +1,6 @@
-# Distributed under the MIT software license, see the accompanying
-# file LICENSE or http://www.opensource.org/licenses/mit-license.php.
-
-"""
-"""
-
-# from OSINTofBlockchain.BitcoinData.BitcoinNodeHelper import BitcoinNodeHelper
-
 # my_node: BitcoinNodeHelper = None
-
-# TODO - MAYBE THESE AREN'T THE BEST.. BUT THEY WORK FOR NOW...
-# price = None
-# height = None
-# difficulty = None
-# hashrate = None
 
 # the number of projections that we've run so we can increment the number when they are displayed
 analysis_number = 0
 
-APP_TITLE = "Treasury Management"
+APP_TITLE = "Treasury Management Simulator"
