+++ web/templates/index.html
@@ -102,14 +102,14 @@
 <div class="row">
   <div class="card">
     <h4 class="card-title">Bitcoin stats dashboard</h5>
-      <img src="/static/bc.jpg" class="card-img-top">
+      <img src="/static/stub.png" class="card-img-top">
       <div class="card-body">
         <a href="/dashboard" class="btn btn-primary">Launch tool</a>
       </div>
     </div>
   <div class="card">
     <h4 class="card-title">Historical analysis</h5>
-      <img src="/static/bc.jpg" class="card-img-top">
+      <img src="/static/stub.png" class="card-img-top">
       <div class="card-body">
         <a href="/historicalanalysis" class="btn btn-primary">Launch tool</a>
       </div>
